pyproject.toml
[tool.poetry]
name = "pyzoom"
version = "0.1.14"
description = "Python wrapper for Zoom Video API"
authors = ["MB <mb@blaster.ai>"]
license = "MIT"
readme='README.md'
homepage='https://github.com/licht1stein/pyzoom'

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.23.0"
attrs = "*"
pyjwt = ">=2.0,<3.0.0"
typing-extensions = "^3.7.4"
pydantic = "^1.5.1"
cachetools = "^4.1.0"
shortuuid = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.4b2"
python-dotenv = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
